pyproject.toml
[project]
name = "ai-chatbot-meetings"
version = "0.2.0"
description = "Streamlit chatbot with tool calling against cal.com API to schedule and update meetings."
keywords = ["chatbot", "streamlit", "open-ai", "langchain", "tool-calling"]
readme = "README.md"
requires-python = ">=3.12"

# When updating dependencies, also do `uv lock` and git commit uv.lock 
dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-community>=0.3,<0.4",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.27",
    "python-dotenv>=1.0.0",
    "streamlit",
    "requests>=2.31.0",
    "streamlit-js-eval",
    "pytz",
]
authors = [
  {name = "Chad Norwood", email = "chad@chadnorwood.com"},
]
maintainers = [
  {name = "Chad Norwood", email = "chad@chadnorwood.com"},
]

[project.urls]
Homepage = "https://ai-chatbot-meetings.streamlit.app"
Repository = "https://github.com/chadn/ai-chatbot-meetings.git"


# Below from:
# https://gist.github.com/GhostofGoes/75051c4aeb215bc3cf48c10f5454b399
#
# Development dependencies
# uv pip install -e .[dev]
# uv pip install -e .[lint,test,exe]
[project.optional-dependencies]
lint = [
    # ruff linter checks for issues and potential bugs
    "ruff",

    # automatic code formatting to follow a consistent style
    "black",
]

test = [
    # pytest for running tests
    "pytest",
    "pytest-asyncio",
]

dev = [
    # Can use https://github.com/cle-b/httpdbg to run streamlit python app
    "httpdbg",
]

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (may require real API keys)"
]
